Prepare the v0.7 by some wtf things
